# api/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Rncasp API
  description: >
    REST API for the Rncasp (Really No Clue About Shift Planning) application.
    A shift planning web application for LAN party events that replaces Excel-based
    workflows where team members sign up for shifts across multi-day events.


    All responses (except health and binary downloads) are wrapped in a standard
    envelope: `{"data": ..., "error": ..., "meta": ...}`.


    Authentication is cookie-based using session cookies set on login/register.
  version: 1.0.0
  contact:
    name: Rncasp
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
...
security:
  - cookieAuth: []
paths:
...
  /api/events/{slug}/coverage:
    get:
      tags: [Coverage]
      operationId: listCoverage
      summary: List coverage requirements for an event
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      responses:
        "200":
          description: List of coverage requirements
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CoverageRequirement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Coverage]
      operationId: createCoverage
      summary: Create a coverage requirement
      description: Event admin or super-admin.
      parameters:
        - $ref: "#/components/parameters/EventSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCoverageRequest"
      responses:
        "201":
          description: Coverage requirement created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverageRequirementResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{slug}/coverage/{coverageId}:
    put:
      tags: [Coverage]
      operationId: updateCoverage
      summary: Update a coverage requirement
      description: Event admin or super-admin.
      parameters:
        - $ref: "#/components/parameters/EventSlug"
        - $ref: "#/components/parameters/CoverageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCoverageRequest"
      responses:
        "200":
          description: Coverage requirement updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverageRequirementResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Coverage]
      operationId: deleteCoverage
      summary: Delete a coverage requirement
      description: Event admin or super-admin.
      parameters:
        - $ref: "#/components/parameters/EventSlug"
        - $ref: "#/components/parameters/CoverageId"
      responses:
        "200":
          description: Coverage requirement deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/events/{slug}/coverage/team/{teamId}:
    delete:
      tags: [Coverage]
      operationId: deleteCoverageByTeam
      summary: Delete all coverage requirements for a team in an event
      description: Event admin or super-admin.
      parameters:
        - $ref: "#/components/parameters/EventSlug"
        - name: teamId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Coverage requirements deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Availability
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: >
        Session cookie set by the login or register endpoints.
        The cookie name is configurable; `session` is the default.

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
    EventSlug:
      name: slug
      in: path
      required: true
      description: URL-friendly event identifier
      schema:
        type: string
        pattern: "^[a-zA-Z0-9_-]+$"
...
    CoverageId:
      name: coverageId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: invalid_input
              message: "invalid request body"
    Unauthorized:
      description: Not authenticated or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: unauthorized
              message: "not authenticated"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: forbidden
              message: "insufficient permissions"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            error:
              code: not_found
              message: "resource not found"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/APIError"
    APIError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - not_found
            - already_exists
            - conflict
            - unauthorized
            - forbidden
            - invalid_input
            - event_locked
            - overbooking_not_allowed
            - account_inactive
            - dummy_account_cannot_login
            - invalid_totp_code
            - totp_required
            - too_many_requests
        message:
          type: string
          description: Human-readable error message
        field:
          type: string
          description: Field name for validation errors
...
    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
          required: [message]

    # --- Health ---
...
    CoverageRequirement:
      type: object
      required: [id, event_id, team_id, start_time, end_time, required_count]
      properties:
        id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        required_count:
          type: integer
    CoverageRequirementResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/CoverageRequirement"
    CreateCoverageRequest:
      type: object
      required: [team_id, start_time, end_time, required_count]
      properties:
        team_id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        required_count:
          type: integer
          minimum: 0
    UpdateCoverageRequest:
      type: object
      required: [team_id, start_time, end_time, required_count]
      properties:
        team_id:
          type: string
          format: uuid
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        required_count:
          type: integer
          minimum: 0

    # --- Availability ---
...